Factura Electrónica 3 (métodos de facturación)

+++ core/application/controllers/cotizaciones.php
@@ -824,43 +824,49 @@ class Cotizaciones extends CI_Controller {
 			$this->mpdf->Output('./tmp/'.$file, 'F');
 
 
-			$this->load->library('email');
+			$this->load->model('facturaelectronica');
+			$email_data = $this->facturaelectronica->get_email();
 
-			$config['protocol']    = 'smtp';
-			$config['smtp_host']    = 'ssl://smtp.gmail.com';
-			$config['smtp_port']    = '465';
-			$config['smtp_timeout'] = '7';
-			$config['smtp_user']    = 'contactodeik@gmail.com';
-			$config['smtp_pass']    = 'deikk1965';
-			$config['charset']    = 'utf-8';
-			$config['newline']    = "\r\n";
-			$config['mailtype'] = 'html'; // or html
-			$config['validation'] = TRUE; // bool whether to validate email or not      			
+			if(count($email_data) > 0){
+				$this->load->library('email');
+				$config['protocol']    = $email_data->tserver_intercambio;
+				$config['smtp_host']    = $email_data->host_intercambio;
+				$config['smtp_port']    = $email_data->port_intercambio;
+				$config['smtp_timeout'] = '7';
+				$config['smtp_user']    = $email_data->email_intercambio;
+				$config['smtp_pass']    = $email_data->pass_intercambio;
+				$config['charset']    = 'utf-8';
+				$config['newline']    = "\r\n";
+				$config['mailtype'] = 'html'; // or html
+				$config['validation'] = TRUE; // bool whether to validate email or not      			
 
-			$this->email->initialize($config);		  		
+				$this->email->initialize($config);		  		
 
-		    $this->email->from('contactodeik@gmail.com', 'Deik');
-		    $this->email->to($email);
+			    $this->email->from($email_data->email_intercambio, NOMBRE_EMPRESA);
+			    $this->email->to($email);
 
-		    //$this->email->bcc(array('rodrigo.gonzalez@info-sys.cl','cesar.moraga@info-sys.cl','sergio.arriagada@info-sys.cl','rene.gonzalez@info-sys.cl')); 
-		    $this->email->subject('Envio de Cotizacion');
-		    $this->email->message($mensaje);
+			    //$this->email->bcc(array('rodrigo.gonzalez@info-sys.cl','cesar.moraga@info-sys.cl','sergio.arriagada@info-sys.cl','rene.gonzalez@info-sys.cl')); 
+			    $this->email->subject('Envio de Cotizacion');
+			    $this->email->message($mensaje);
 
-		    $this->email->attach('./tmp/'.$file,'attachment', 'Cotizacion.pdf');		
-		    
-		    try {
-		      $this->email->send();
-		      //var_dump($this->email->print_debugger());
-		      unlink('./tmp/'.$file);
-		      	        exit;
-		    } catch (Exception $e) {
-		      echo $e->getMessage() . '<br />';
-		      echo $e->getCode() . '<br />';
-		      echo $e->getFile() . '<br />';
-		      echo $e->getTraceAsString() . '<br />';
-		      echo "no";
+			    $this->email->attach('./tmp/'.$file,'attachment', 'Cotizacion.pdf');			
 
+
+			    try {
+			      $this->email->send();
+			      //var_dump($this->email->print_debugger()); exit;
+			      unlink('./tmp/'.$file);
+			      	        exit;
+			    } catch (Exception $e) {
+			      echo $e->getMessage() . '<br />';
+			      echo $e->getCode() . '<br />';
+			      echo $e->getFile() . '<br />';
+			      echo $e->getTraceAsString() . '<br />';
+			      echo "no";
+
+			    }
 		    }
+		    
 		exit;
 	}
 }
